Forcefully hide the <video> when offline.
Without `!important`, the `.player.is-full-screen video` selector
would override `.player:not([data-status="playing"]) video`,
setting `display` to `block` instead of `none`. If the screen
has lower vertical resolution than the video, the blank svg
will not cover the whole element, leaving a small part of the
stopped video visible at the bottom. Which is bad.