Allow to specify a callback to run when a stream is closed.
This will be used to implement stream balancing.
Each stream will be owned by a single retransmission node
which would demux it, forward it, handle the chat, etc.
The fact of the ownership will be recorded in the database
when the stream is created, kept while the stream is inactive,
and erased when it is closed for good.