Always run in joint mode.
Also, use server-side redirects to point clients to the correct node.

The idea is that all nodes have unique public addresses, and
a common domain name. Thus basic DNS round-robin could be used
to balance requests between nodes, but viewers may be redirected
to a different server if desired. (For example, when trying to view
a stream that is being broadcast from a different node, or when
trying to broadcast through an already overloaded node.)