:tada: finally implement an emoji picker.
Obviously, the most important part of any webscale app. /s

It's kinda slow, though. (There are ~1300 emojis in latest Unicode,
and not a single one that I could use instead of that "/s".)
Oh, and the list of categories does not fit into the chat horizontally.
(Flexbox items only shrink when you don't want them to, I guess.)